+++ app/Http/Controllers/Admin/ProductController.php
@@ -104,6 +104,7 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        abort_if(request()->user()->role_id, 'Not Allowed.');
         $product->delete();
 
         return request()->ajax()
